update
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@26477 52c4871e-980c-0410-b1e0-e73912ce01f8

Former-commit-id: f186607c70f9268395aa9d0555e4196846cc98c7

+++ src/router/quagga/ospf6d/ospf6_spf.c
@@ -40,6 +40,7 @@
 #include "ospf6_intra.h"
 #include "ospf6_interface.h"
 #include "ospf6d.h"
+#include "ospf6_abr.h"
 
 unsigned char conf_debug_ospf6_spf = 0;
 
@@ -391,7 +392,7 @@ static const char *ospf6_spf_reason_str[] =
 
 void ospf6_spf_reason_string (unsigned int reason, char *buf, int size)
 {
-  int bit;
+  size_t bit;
   int len = 0;
 
   if (!buf)
